+++ LibOS/shim/src/sys/shim_time.c
@@ -25,9 +25,9 @@ int shim_do_gettimeofday(struct __kernel_timeval* tv, struct __kernel_timezone* 
     if (tz && test_user_memory(tz, sizeof(*tz), true))
         return -EFAULT;
 
-    long time = DkSystemTimeQuery();
+    uint64_t time = DkSystemTimeQuery();
 
-    if (time == -1)
+    if (time == (uint64_t)-1)
         return -PAL_ERRNO();
 
     tv->tv_sec  = time / 1000000;
@@ -36,9 +36,9 @@ int shim_do_gettimeofday(struct __kernel_timeval* tv, struct __kernel_timezone* 
 }
 
 time_t shim_do_time(time_t* tloc) {
-    long time = DkSystemTimeQuery();
+    uint64_t time = DkSystemTimeQuery();
 
-    if (time == -1)
+    if (time == (uint64_t)-1)
         return -PAL_ERRNO();
 
     if (tloc && test_user_memory(tloc, sizeof(*tloc), true))
@@ -62,9 +62,9 @@ int shim_do_clock_gettime(clockid_t which_clock, struct timespec* tp) {
     if (test_user_memory(tp, sizeof(*tp), true))
         return -EFAULT;
 
-    long time = DkSystemTimeQuery();
+    uint64_t time = DkSystemTimeQuery();
 
-    if (time == -1)
+    if (time == (uint64_t)-1)
         return -PAL_ERRNO();
 
     tp->tv_sec  = time / 1000000;
